Ignore DDM chunk commands
Honestly, I have no idea why, but ADB occasionally returns DDM chunk
chunk commands for commands for which we previously received a response.
These commands are characterized by the commandset 199 (which is in the
custom vendor range of 128-255) and a command of 1.

Because these responses refer to commands for which we've already
received a response, the code actually blows up entirely with a (hidden)
null reference exception in the error printing code (it attempts to
dereference this.command which is now undefined).

This fix adds special handling for these commands, though nothing
sophisticated: it just ignores them.

In addition, we fix up the error handling code so that if multiple
responses for a command are received in the future, we don't blow up.